+++ pyproject.toml
@@ -1,3 +1,12 @@
+[project]
+name = "llm-open-data-insee"
+version = "0.1.0"
+description = "Add your description here"
+readme = "README.md"
+requires-python = ">=3.12"
+dependencies = []
+
+
 [tool.ruff]
 line-length = 120
 indent-width = 4
